+++ test/Types/AttachmentTest.php
@@ -18,7 +18,7 @@ class AttachmentTest extends \PHPUnit_Framework_TestCase
          * Use a class that will fake sending requests and getting responses.
          * The idea is that all the information needed to make the request is
          * passed to the client by the service. What we want to test is that the
-         * is actualy passed correctly. We are not testing the sending of the request
+         * is actually passed correctly. We are not testing the sending of the request
          * over the internet.
          * The HttpHandler contains properties that will be set when the service
          * makes the request. We can test these properties to check what the service is passing.